feat: add admin checker

+++ main.py
@@ -22,7 +22,7 @@ async def on_callback_query(event):
 async def handler(event):
     user = await event.get_user()
     if event.user_joined:
-        await event.reply(f"Welcome {user.username} to the Myanmar Cyber Youths development team 🐙")
+        await event.reply(f"🐙 Welcome {user.username} to the Myanmar Cyber Youths")
 
 
 @client.on(events.NewMessage(pattern='/start'))
@@ -43,8 +43,17 @@ async def on_new_message(event):
 
 @client.on(events.NewMessage(pattern='/leave'))
 async def handle_start_command(event):
+    print(f"Received a message from {event.sender.username}: {event.message.message}")
+
     sender = await event.get_sender()
     chat_id = event.message.peer_id
+    permission = await client.get_permissions(chat_id, sender.id)
+
+    if (not permission.is_admin and not permission.is_creator and not config.get(
+        'TELEGRAM_ADMIN_USERNAME') == sender.username):
+        await event.reply('😾 You are not allowed to use this command')
+        return
+
     if isinstance(chat_id, (PeerUser, PeerChat, PeerChannel)):
         await client.send_message(chat_id, 'Leaving the chat')
         await client.delete_dialog(chat_id)
